fix(crm): complete stage-probability import move (3 importers missed in prior commit)
The L1 module-move commit captured only the rename + crm-tools.ts; compute.ts,
stage-style.ts, and crm-tools.test.ts still imported @/server/crm/stage-probability
(now moved to @/lib/crm) — the pushed branch would fail tsc/build/L1. Complete the
import sweep so all importers resolve.

Ref #6172
