+++ apps/web-platform/components/crm/stage-style.ts
@@ -1,12 +1,12 @@
 // UI styling for the beta-CRM pipeline stages (feat-beta-crm-ui #6172). The
 // ORDERED stage enum is imported from the single source of truth
-// (server/crm/stage-probability.ts) and never re-declared here (AC7) — these
+// (lib/crm/stage-probability.ts) and never re-declared here (AC7) — these
 // maps are keyed by stage (unquoted identifiers, not string literals) so a
 // stage-list grep finds nothing, and Record<Stage, …> makes a missing stage a
 // COMPILE error (a free drift guard). Accent hexes mirror the operator-approved
 // wireframe (knowledge-base/product/design/crm/beta-crm-pipeline.pen).
 
-import { STAGES, FUNNEL_STAGES, type Stage } from "@/server/crm/stage-probability";
+import { STAGES, FUNNEL_STAGES, type Stage } from "@/lib/crm/stage-probability";
 
 export { STAGES, FUNNEL_STAGES };
 export type { Stage };
